@@ -27,7 +27,7 @@ environment:
 # the latest version available on pub.dev. To see which dependencies have newer
 # versions available, run `flutter pub outdated`.
 dependencies:
-  smooth_page_indicator: ^1.0.1
+  smooth_page_indicator: ^1.1.0
   flutter:
     sdk: flutter
 
@@ -37,26 +37,26 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
   font_awesome_flutter: ^10.7.0
-  url_launcher: ^6.1.11
-  firebase_core: ^2.14.0
-  firebase_storage: ^11.1.1
-  file_picker: ^5.2.9
-  firebase_auth: ^4.6.3
-  firebase_database: ^10.1.1
-  google_sign_in: ^6.1.4
-  flutter_polyline_points: ^1.0.0
-  google_maps_flutter: ^2.2.5
-  location: ^4.4.0
-  geolocator: ^9.0.2
-  cloud_firestore: ^4.7.0
+  url_launcher: ^6.2.6
+  firebase_core: ^2.32.0
+  firebase_storage: ^11.7.7
+  file_picker: ^8.0.3
+  firebase_auth: ^4.20.0
+  firebase_database: ^10.5.7
+  google_sign_in: ^6.2.1
+  flutter_polyline_points: ^2.0.0
+  google_maps_flutter: ^2.6.1
+  location: ^6.0.2
+  geolocator: ^12.0.0
+  cloud_firestore: ^4.17.5
   flutter_rating_bar: ^4.0.1
   flutter_pannable_rating_bar: ^2.7.2+1
-  app_settings: ^4.2.0
-  flutter_localizations:
-    sdk: flutter
-  intl: ^0.18.0
-  shared_preferences: ^2.1.2
-  provider:
+  app_settings: ^5.1.1
+  intl: ^0.19.0
+  shared_preferences: ^2.2.3
+  provider: ^6.1.2
+  easy_localization: ^3.0.7
+  google_mobile_ads: ^5.1.0
 
 
 
@@ -70,7 +70,7 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^1.0.0
+  flutter_lints: ^4.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
@@ -111,6 +111,7 @@ flutter:
     - images/derekoy/
     - images/sirinkoy/
     - images/atm/
+    - assets/translations/
 
 
   # An image asset can refer to one or more resolution-specific "variants", see
@@ -127,13 +128,13 @@ flutter:
   fonts:
     - family: Montserrat
       fonts:
-        - asset: lib/assets/fonts/Montserrat-Regular.ttf
+        - asset: assets/fonts/Montserrat-Regular.ttf
     #       - asset: fonts/Schyler-Italic.ttf
     #         style: italic
     - family: Poppins
       fonts:
-        - asset: lib/assets/fonts/Poppins-Regular.ttf
-        - asset: lib/assets/fonts/Poppins-Bold.ttf
+        - asset: assets/fonts/Poppins-Regular.ttf
+        - asset: assets/fonts/Poppins-Bold.ttf
   #         weight: 700
   #
   # For details regarding fonts from package dependencies,
