@@ -4,15 +4,15 @@ class ColorConstants {
   static ColorConstants instance = ColorConstants._init();
   ColorConstants._init();
 
-  final activatedButton = Color(0xffee6723);
-  final roomNumber = Color(0xffee6723);
-  final starColor = Color(0xfffdbd39);
-  final titleColor = Color(0xff554d56);
-  final bottomBarIcon = Color(0xff554d56);
-  final commentColor = Color(0xff979197);
-  final searcButtonColor = Color(0xffc1bec1);
-  final textFieldTextColor = Color(0xffc1bec1);
-  final lightGreyCardCollor = Color(0xffedeced);
-  final textFieldBacgroundColor = Color(0xfff7f7f7);
-  final imageFrontTextColor = Color(0xffffffff);
+  final activatedButton = const Color(0xffee6723);
+  final roomNumber = const Color(0xffee6723);
+  final starColor = const Color(0xfffdbd39);
+  final titleColor = const Color(0xff554d56);
+  final bottomBarIcon = const Color(0xff554d56);
+  final commentColor = const Color(0xff979197);
+  final searcButtonColor = const Color(0xffc1bec1);
+  final textFieldTextColor = const Color(0xffc1bec1);
+  final lightGreyCardCollor = const Color(0xffedeced);
+  final textFieldBacgroundColor = const Color(0xfff7f7f7);
+  final imageFrontTextColor = const Color(0xffffffff);
 }
