@@ -1,8 +1,9 @@
+import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:gokceada/product/plaj_inside_design.dart';
 import 'package:gokceada/screens/restaurantDetail.dart';
 import '../product/navigationButton.dart';
-import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+
 
 
 Map<String,Widget> createKoyler(BuildContext context) {
@@ -22,9 +23,7 @@ Map<String,Widget> createKoyler(BuildContext context) {
         ],
         widget: const NavigationButton(
             latitude: 40.22132946119443, longitude: 25.90695936121934),
-        description: AppLocalizations
-            .of(context)!
-            .bademli),
+        description: 'bademli'.tr()),
 
     'derekoy': DescriptionCustom(
         title: 'Dereköy',
@@ -39,9 +38,7 @@ Map<String,Widget> createKoyler(BuildContext context) {
         ],
         widget: const NavigationButton(
             latitude: 40.156565879596876, longitude: 25.779833059383733),
-        description: AppLocalizations
-            .of(context)!
-            .derekoy),
+        description: 'derekoy'.tr()),
 
     'eselek': DescriptionCustom(
         title: 'Eşelek Köyü',
@@ -52,7 +49,7 @@ Map<String,Widget> createKoyler(BuildContext context) {
         ],
         widget: const NavigationButton(
             latitude: 40.14759580493827, longitude: 25.94710513716375),
-        description: AppLocalizations.of(context)!.eselek),
+        description: 'eselek'.tr()),
 
     'kalekoy': DescriptionCustom(
         title: 'Kaleköy - Kastro',
@@ -66,7 +63,7 @@ Map<String,Widget> createKoyler(BuildContext context) {
         ],
         widget: const NavigationButton(
             latitude: 40.234366399996055, longitude: 25.89884971836279),
-        description: AppLocalizations.of(context)!.kalekoy),
+        description: 'kalekoy'.tr()),
 
     'kalekoyLiman': DescriptionCustom(
         title: 'Kaleköy Liman',
@@ -84,7 +81,7 @@ Map<String,Widget> createKoyler(BuildContext context) {
         ],
         widget: const NavigationButton(
             latitude: 40.23136718323504, longitude: 25.894804669123182),
-        description: AppLocalizations.of(context)!.kalekoyLiman),
+        description: 'kalekoyLiman'.tr()),
 
 
     'merkez': DescriptionCustom(
@@ -98,7 +95,7 @@ Map<String,Widget> createKoyler(BuildContext context) {
         ],
         widget: const NavigationButton(
             latitude: 40.19624748335972, longitude: 25.90766888323214),
-        description: AppLocalizations.of(context)!.merkez),
+        description: 'merkez'.tr()),
 
     'sahinkaya': DescriptionCustom(
         title: 'Şahinkaya Köyü',
@@ -107,7 +104,7 @@ Map<String,Widget> createKoyler(BuildContext context) {
         ],
         widget: const NavigationButton(
             latitude: 40.150813785785985, longitude: 25.785779654393263),
-        description: AppLocalizations.of(context)!.sahinkaya),
+        description: 'sahinkaya'.tr()),
 
     'sirinkoy': DescriptionCustom(
         title: 'Şirinköy',
@@ -117,7 +114,7 @@ Map<String,Widget> createKoyler(BuildContext context) {
         ],
         widget: const NavigationButton(
             latitude: 40.1286817177055, longitude: 25.74225144997795),
-        description: AppLocalizations.of(context)!.sirinkoy),
+        description: 'sirinkoy'.tr()),
 
     'tepekoy': DescriptionCustom(
         title: 'Tepeköy - Agridia',
@@ -133,7 +130,7 @@ Map<String,Widget> createKoyler(BuildContext context) {
         ],
         widget: const NavigationButton(
             latitude: 40.14860030447957, longitude: 25.776397678636698),
-        description: AppLocalizations.of(context)!.tepekoy),
+        description: 'tepekoy'.tr()),
 
     'ugurlu': DescriptionCustom(
         title: 'Uğurlu Köyü',
@@ -144,7 +141,7 @@ Map<String,Widget> createKoyler(BuildContext context) {
         ],
         widget: const NavigationButton(
             latitude: 40.12373775008842, longitude: 25.70627252370199),
-        description: AppLocalizations.of(context)!.ugurlu),
+        description: 'ugurlu'.tr()),
 
     'yenibademli': DescriptionCustom(
         title: 'Yenibademli',
@@ -155,7 +152,7 @@ Map<String,Widget> createKoyler(BuildContext context) {
         ],
         widget: const NavigationButton(
             latitude: 40.229279376684836, longitude: 25.906215449547606),
-        description: AppLocalizations.of(context)!.yenibademli),
+        description: 'yenibademli'.tr()),
 
     'zeytinlikoy': DescriptionCustom(
         title: 'Zeytinliköy - Aya Teodoroi',
@@ -169,7 +166,7 @@ Map<String,Widget> createKoyler(BuildContext context) {
         ],
         widget: const NavigationButton(
             latitude: 40.19042836024782, longitude: 25.871124776652493),
-        description: AppLocalizations.of(context)!.zeytinlikoy)
+        description: 'zeytinlikoy'.tr())
   };
 }
 
