@@ -1,8 +1,8 @@
+import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import '../../core/colors.dart';
 import '../../core/textFont.dart';
 import '../../product/card_design.dart';
-import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 class NerdeYenirView extends StatelessWidget {
   const NerdeYenirView({Key? key}) : super(key: key);
@@ -13,17 +13,17 @@ class NerdeYenirView extends StatelessWidget {
       appBar: AppBar(
       backgroundColor: Colors.white,
       foregroundColor: ColorConstants.instance.titleColor,
-      title: Text(AppLocalizations.of(context)!.neredeYenir, style: TextFonts.instance.appBarTitleColor),
+      title: Text('neredeYenir'.tr(), style: TextFonts.instance.appBarTitleColor),
       elevation: 0.8,
       ),
       body:Padding(
         padding: const EdgeInsets.all(8.0),
         child: ListView(
           children:  [
-            CardDesign(path: 'images/nerdeYenir/poseidon2.jpg', cardText: AppLocalizations.of(context)!.restoranlar, pushWhere: 'restaurantsView'),
-            CardDesign(path: 'images/nerdeYenir/bar.jpg', cardText: AppLocalizations.of(context)!.barlar, pushWhere: 'bar'),
-            CardDesign(path: 'images/nerdeYenir/cafe.jpg', cardText: AppLocalizations.of(context)!.kafeler, pushWhere: 'cafe'),
-            CardDesign(path: 'images/nerdeYenir/kahvalti.jpg', cardText: AppLocalizations.of(context)!.kahvaltiYerleri, pushWhere: 'kahvalti'),
+            CardDesign(path: 'images/nerdeYenir/poseidon2.jpg', cardText: 'restoranlar'.tr(), pushWhere: 'restaurantsView'),
+            CardDesign(path: 'images/nerdeYenir/bar.jpg', cardText: 'barlar'.tr(), pushWhere: 'bar'),
+            CardDesign(path: 'images/nerdeYenir/cafe.jpg', cardText: 'kafeler'.tr(), pushWhere: 'cafe'),
+            CardDesign(path: 'images/nerdeYenir/kahvalti.jpg', cardText: 'kahvaltiYerleri'.tr(), pushWhere: 'kahvalti'),
           ],
         ),
       ),
