@@ -1,3 +1,4 @@
+import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:gokceada/core/ratingBar.dart';
 import 'package:gokceada/product/hotelListCard.dart';
@@ -5,7 +6,6 @@ import 'package:gokceada/screens/hotel_rooms.dart';
 import 'package:url_launcher/url_launcher.dart';
 import '../core/textFont.dart';
 import '../product/navigationButton.dart';
-import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 class RestaurantView extends StatefulWidget {
   const RestaurantView(
@@ -91,11 +91,11 @@ class _RestaurantViewState extends State<RestaurantView> {
               ),
               const SizedBox(height: 30),
               Text(
-                  AppLocalizations.of(context)!.menuBilgilendirme,
+                  'menuBilgilendirme'.tr(),
                   style: TextFonts.instance.commentTextBold),
               const SizedBox(height: 8),
               InkwellUnderline(
-                  name: 'Web Site',
+                  name: 'QR Menu',
                   onTap: () {
                     launchUrl(url);
                   }),
