@@ -1,10 +1,11 @@
+import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:gokceada/product/restaurantsCard.dart';
 import 'package:gokceada/screens/restaurantDetail.dart';
 import '../core/colors.dart';
 import '../core/textFont.dart';
-import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+
 
 class RestaurantDetay extends StatefulWidget {
   const RestaurantDetay({Key? key}) : super(key: key);
@@ -71,7 +72,7 @@ class _RestaurantDetayState extends State<RestaurantDetay> {
           },
           icon: Icon(Icons.arrow_back_ios_new, color: ColorConstants.instance.titleColor),
         ),
-        title: Text(AppLocalizations.of(context)!.restoranlar, style: TextFonts.instance.titleFont),
+        title: Text('restoranlar'.tr(), style: TextFonts.instance.titleFont),
       ),
       body: ListView.builder(
         itemCount: restaurants.length,
